Bias-Variance Commit
In every predictive model, there are some reducible errors associated with it. They are Bias and Variance. Bias leads to underfitting and Variance leads to overfitting which ultimately leads to high Test error. We can avoid them by using Regularization in our model. This notebook  contains some experiment to see the effect of Bias and  Variance and how to avoid it.